// src/http-courses/models/empregabilidadeStatusCandidatura.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * API Go
 * API de serviços para aplicativos da Prefeitura do Rio
 * OpenAPI spec version: 1.0
 */

export type EmpregabilidadeStatusCandidatura =
  (typeof EmpregabilidadeStatusCandidatura)[keyof typeof EmpregabilidadeStatusCandidatura]

export const EmpregabilidadeStatusCandidatura = {
  StatusCandidaturaEnviada: 'candidatura_enviada',
  StatusCandidaturaAprovada: 'aprovada',
  StatusCandidaturaReprovada: 'reprovada',
  StatusCandidaturaVagaCongelada: 'vaga_congelada',
  StatusCandidaturaDescontinuada: 'vaga_descontinuada',
} as const